requirements.txt
Django==5.2.10
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0